backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.10

# File handling
python-multipart==0.0.12
pillow==10.4.0
aiofiles==24.1.0

# Astronomy calculations
astropy==6.1.4
scipy==1.14.1
numpy==1.26.4

# Configuration
python-dotenv==1.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
email-validator==2.1.0
